+++ e2e/BinaryCoStream/src/lib/getDownloaderPeerUrl.ts
@@ -1,7 +1,0 @@
-import { UploadedFile } from "src/schema";
-
-export function getDownloaderPeerUrl(value: UploadedFile) {
-  const url = new URL(window.location.href);
-  url.searchParams.set("valueId", value.id);
-  return url.toString();
-}