+++ homepage/homepage/app/examples/page.tsx
@@ -2,11 +2,12 @@ import { ClerkFullLogo } from "@/components/icons/ClerkFullLogo";
 import { NextjsLogo } from "@/components/icons/NextjsLogo";
 import { ReactLogo } from "@/components/icons/ReactLogo";
 import { ReactNativeLogo } from "@/components/icons/ReactNativeLogo";
+import { VueLogo } from "@/components/icons/VueLogo";
 import { Button } from "gcmp-design-system/src/app/components/atoms/Button";
 import { H2 } from "gcmp-design-system/src/app/components/atoms/Headings";
 import { GappedGrid } from "gcmp-design-system/src/app/components/molecules/GappedGrid";
 import { HeroHeader } from "gcmp-design-system/src/app/components/molecules/HeroHeader";
-import { CloudUploadIcon, FingerprintIcon, KeyRoundIcon } from "lucide-react";
+import { CloudUploadIcon, FingerprintIcon, ImageIcon } from "lucide-react";
 
 type Example = {
   name: string;
@@ -22,27 +23,36 @@ const tech = {
   react: "React",
   nextjs: "Next.js",
   reactNative: "React Native",
+  vue: "Vue",
 };
 
 const features = {
   fileUpload: "File upload",
+  imageUpload: "Image upload",
   passkey: "Passkey auth",
   clerk: "Clerk auth",
   inviteLink: "Invite link",
+  coFeed: "CoFeed",
 };
+
+const MockButton = ({ children }: { children: React.ReactNode }) => (
+  <p className="bg-blue-100 text-blue-800 py-1 p-2 rounded-full font-medium text-center text-xs">
+    {children}
+  </p>
+);
 
 const ChatIllustration = () => (
   <div className="p-4 flex flex-col gap-4 justify-center h-full">
     <div className="flex flex-col gap-1 items-end">
       <p className="text-2xs">Sebastian</p>
-      <p className="inline-block text-xs py-1.5 px-3 rounded-full bg-blue text-white">
+      <p className="inline-block text-xs py-1.5 p-2 rounded-full bg-blue text-white">
         No one likes jazz. Not even you.
       </p>
     </div>
 
     <div className="flex flex-col gap-1 items-start">
       <p className="text-2xs">Mia</p>
-      <p className="inline-block text-xs py-1.5 px-3 rounded-full bg-stone-200 text-stone-900 dark:bg-white">
+      <p className="inline-block text-xs py-1.5 p-2 rounded-full bg-stone-200 text-stone-900 dark:bg-white">
         I do like jazz now, because of you.
       </p>
     </div>
@@ -57,7 +67,7 @@ const ClerkIllustration = () => (
 
 const MusicIllustration = () => (
   <div className="flex flex-col items-center justify-center h-full p-8">
-    <div className="py-3 px-8 border border-dashed border-blue dark:border-blue-500 rounded-lg flex gap-2 flex-col items-center">
+    <div className="p-3 w-[12rem] h-[8rem] border border-dashed border-blue dark:border-blue-500 rounded-lg flex gap-2 flex-col items-center justify-center">
       <CloudUploadIcon
         size={40}
         strokeWidth={1.5}
@@ -70,6 +80,37 @@ const MusicIllustration = () => (
   </div>
 );
 
+const ImageUploadIllustration = () => (
+  <div className="flex flex-col items-center justify-center h-full p-8">
+    <div className="p-3 w-[12rem] h-[8rem] border border-dashed border-blue dark:border-blue-500 rounded-lg flex gap-2 flex-col items-center justify-center">
+      <ImageIcon
+        size={40}
+        strokeWidth={1.5}
+        className="stroke-blue mx-auto dark:stroke-blue-500"
+      />
+      <p className="whitespace-nowrap text-stone-900 dark:text-white">
+        profile-photo.jpg
+      </p>
+    </div>
+  </div>
+);
+
+const ReactionsIllustration = () => (
+  <div className="flex bg-stone-100 h-full flex-col items-center justify-center dark:bg-transparent">
+    <div className="inline-flex justify-center gap-1.5 mx-auto">
+      {["😍", "😮", "🤩", "😂", "👍"].map((emoji) => (
+        <button
+          type="button"
+          key={emoji}
+          className="size-10 text-xl rounded shadow-sm bg-white leading-none"
+        >
+          {emoji}
+        </button>
+      ))}
+    </div>
+  </div>
+);
+
 const BookShelfIllustration = () => (
   <div className="h-full max-w-[30rem] mx-auto flex flex-col justify-center p-6 md:p-5">
     <div className="flex justify-between items-baseline">
@@ -77,9 +118,7 @@ const BookShelfIllustration = () => (
         Your book shelf
       </p>
 
-      <p className="bg-blue-100 text-blue-800 py-1 px-3 rounded-full font-medium text-xs">
-        Add book
-      </p>
+      <MockButton>Add book</MockButton>
     </div>
 
     <div className="grid grid-cols-3 gap-4 mt-3">
@@ -102,12 +141,68 @@ const PetIllustration = () => (
         <button
           type="button"
           key={emoji}
-          className="size-6 rounded shadow-sm bg-white leading-none hover:bg-stone-100"
+          className="size-6 rounded shadow-sm bg-white leading-none"
         >
           {emoji}
         </button>
       ))}
     </div>
+  </div>
+);
+
+const PasswordManagerIllustration = () => (
+  <div className="max-w-[30rem] mx-auto flex flex-col justify-center h-full p-5 gap-4">
+    <div className="flex justify-between items-center">
+      <p className="font-display font-medium tracking-tight text-sm text-stone-900 dark:text-white">
+        Password manager
+      </p>
+
+      <button
+        type="button"
+        className="border py-1 p-2 rounded-full font-medium text-xs"
+      >
+        Log out
+      </button>
+    </div>
+
+    <table className="text-xs">
+      <thead>
+        <tr className="w-full text-stone-700 bg-stone-50 border-b dark:bg-transparent dark:text-stone-400">
+          <th className="font-medium tracking-wider text-left uppercase p-2">
+            Username
+          </th>
+          <th className="font-medium tracking-wider text-left uppercase p-2">
+            URI
+          </th>
+          <th className="font-medium tracking-wider text-left uppercase p-2">
+            Actions
+          </th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr className="border-b">
+          <td className="p-2">user@gmail.com</td>
+          <td className="p-2">gmail.com</td>
+          <td className="p-2">
+            <MockButton>Copy password</MockButton>
+          </td>
+        </tr>
+        <tr className="border-b">
+          <td className="p-2">user@gmail.com</td>
+          <td className="p-2">fb.com</td>
+          <td className="p-2">
+            <MockButton>Copy password</MockButton>
+          </td>
+        </tr>
+        <tr className="border-b">
+          <td className="p-2">user@gmail.com</td>
+          <td className="p-2">x.com</td>
+          <td className="p-2">
+            <MockButton>Copy password</MockButton>
+          </td>
+        </tr>
+      </tbody>
+    </table>
   </div>
 );
 
@@ -121,24 +216,51 @@ const reactExamples = [
     illustration: <ChatIllustration />,
   },
   {
-    name: "Chat",
-    slug: "chat-clerk",
-    description:
-      "Exactly like the chat app, but it uses Clerk for authentication.",
+    name: "Clerk",
+    slug: "clerk",
+    description: "A React app that uses Clerk for authentication",
     tech: [tech.react],
     features: [features.clerk],
-    demoUrl: "https://chat-clerk-demo.jazz.tools",
+    demoUrl: "https://clerk-demo.jazz.tools",
     illustration: <ClerkIllustration />,
   },
   {
-    name: "Music player",
-    slug: "music-player",
-    description:
-      "Upload your favorite songs, and share them with your friends.",
+    name: "Passkey",
+    slug: "passkey",
+    description: "A React app that uses Passkey for authentication",
     tech: [tech.react],
-    features: [features.fileUpload],
-    demoUrl: "https://music-demo.jazz.tools",
-    illustration: <MusicIllustration />,
+    features: [features.passkey],
+    demoUrl: "https://passkey-demo.jazz.tools",
+    illustration: (
+      <div className="flex bg-stone-100 h-full flex-col items-center justify-center dark:bg-transparent">
+        <div className="p-4 flex flex-col items-center gap-3 rounded-md shadow-xl shadow-stone-400/20 bg-white dark:shadow-none">
+          <FingerprintIcon
+            size={36}
+            strokeWidth={0.75}
+            className="stroke-red-600"
+          />
+          <p className="text-xs dark:text-stone-900">Continue with Touch ID</p>
+        </div>
+      </div>
+    ),
+  },
+  {
+    name: "Image upload",
+    slug: "image-upload",
+    description: "Learn how to upload and delete images",
+    tech: [tech.react],
+    features: [features.imageUpload],
+    demoUrl: "https://image-upload-demo.jazz.tools",
+    illustration: <ImageUploadIllustration />,
+  },
+  {
+    name: "Reactions",
+    slug: "reactions",
+    description: "Collect and render reactions from multiple users.",
+    tech: [tech.react],
+    features: [features.coFeed],
+    demoUrl: "https://reactions-demo.jazz.tools",
+    illustration: <ReactionsIllustration />,
   },
   {
     name: "Rate my pet",
@@ -146,7 +268,7 @@ const reactExamples = [
     description:
       "Upload a photo of your pet, and invite your friends to react to it.",
     tech: [tech.react],
-    features: [features.fileUpload, features.inviteLink],
+    features: [features.imageUpload, features.inviteLink],
     demoUrl: "https://pets-demo.jazz.tools",
     illustration: <PetIllustration />,
   },
@@ -168,18 +290,17 @@ const reactExamples = [
     tech: [tech.react],
     features: [features.passkey],
     demoUrl: "https://passwords-demo.jazz.tools",
-    illustration: (
-      <div className="flex bg-stone-100 h-full flex-col items-center justify-center dark:bg-transparent">
-        <div className="p-4 flex flex-col items-center gap-3 rounded-md shadow-xl shadow-stone-400/20 bg-white dark:shadow-none">
-          <FingerprintIcon
-            size={36}
-            strokeWidth={0.75}
-            className="stroke-red-600"
-          />
-          <p className="text-xs dark:text-stone-900">Continue with Touch ID</p>
-        </div>
-      </div>
-    ),
+    illustration: <PasswordManagerIllustration />,
+  },
+  {
+    name: "Music player",
+    slug: "music-player",
+    description:
+      "Upload your favorite songs, and share them with your friends.",
+    tech: [tech.react],
+    features: [features.fileUpload],
+    demoUrl: "https://music-demo.jazz.tools",
+    illustration: <MusicIllustration />,
   },
 ];
 
@@ -190,7 +311,7 @@ const nextExamples = [
     description:
       "Track and rate the books you read, readable by everyone with the link.",
     tech: [tech.nextjs],
-    features: [features.fileUpload],
+    features: [features.imageUpload],
     demoUrl: "https://books-demo.jazz.tools",
     imageUrl: "/books.jpg",
     illustration: <BookShelfIllustration />,
@@ -215,6 +336,28 @@ const rnExamples: Example[] = [
     illustration: <ClerkIllustration />,
   },
 ];
+
+const vueExamples: Example[] = [
+  {
+    name: "Chat",
+    slug: "chat-vue",
+    description: "A simple app that creates a chat room with a shareable link.",
+    tech: [tech.vue],
+    illustration: <ChatIllustration />,
+  },
+  {
+    name: "Todo list",
+    slug: "todo-vue",
+    description: "A todo list where you can collaborate with invited guests.",
+    tech: [tech.vue],
+    features: [features.inviteLink],
+    demoUrl: "https://todo-demo.jazz.tools",
+    illustration: (
+      <div className="h-full w-full bg-cover bg-[url('/todo.jpg')] bg-left-bottom"></div>
+    ),
+  },
+];
+
 const categories = [
   {
     name: "React",
@@ -231,6 +374,11 @@ const categories = [
     logo: ReactNativeLogo,
     examples: rnExamples,
   },
+  {
+    name: "Vue",
+    logo: VueLogo,
+    examples: vueExamples,
+  },
 ];
 
 function Example({ example }: { example: Example }) {
@@ -239,12 +387,12 @@ function Example({ example }: { example: Example }) {
   const githubUrl = `https://github.com/gardencmp/jazz/tree/main/examples/${slug}`;
 
   return (
-    <div className="col-span-2 border bg-stone-50 shadow-sm p-3 flex flex-col gap-3 rounded-lg md:gap-4 dark:bg-stone-950">
-      <div className="aspect-[16/9] overflow-hidden w-full rounded-md bg-white border dark:bg-stone-925 sm:aspect-[2/1] md:aspect-[3/2]">
+    <div className="col-span-2 border bg-stone-50 shadow-sm p-3 flex flex-col rounded-lg dark:bg-stone-950">
+      <div className="mb-3 aspect-[16/9] overflow-hidden w-full rounded-md bg-white border dark:bg-stone-925 sm:aspect-[2/1] md:aspect-[3/2]">
         {illustration}
       </div>
 
-      <div className="flex-1 space-y-2">
+      <div className="flex-1 space-y-2 mb-2">
         <h2 className="font-medium text-stone-900 dark:text-white leading-none">
           {name}
         </h2>
@@ -267,16 +415,16 @@ function Example({ example }: { example: Example }) {
           ))}
         </div>
         <p className="text-sm">{description}</p>
-        <div className="flex gap-2">
-          <Button href={githubUrl} variant="secondary" size="sm">
-            View code
+      </div>
+      <div className="flex gap-2">
+        <Button href={githubUrl} variant="secondary" size="sm">
+          View code
+        </Button>
+        {demoUrl && (
+          <Button href={demoUrl} variant="secondary" size="sm">
+            View demo
           </Button>
-          {demoUrl && (
-            <Button href={demoUrl} variant="secondary" size="sm">
-              View demo
-            </Button>
-          )}
-        </div>
+        )}
       </div>
     </div>
   );