+++ packages/cojson/src/exports.ts
@@ -53,7 +53,11 @@ import type {
   Peer,
   SyncMessage,
 } from "./sync.js";
-import { DisconnectedError, PingTimeoutError } from "./sync.js";
+import {
+  DisconnectedError,
+  PingTimeoutError,
+  emptyKnownState,
+} from "./sync.js";
 
 type Value = JsonValue | AnyRawCoValue;
 
@@ -116,6 +120,7 @@ export {
   SyncMessage,
   isRawCoID,
   LSMStorage,
+  emptyKnownState,
 };
 
 export type {
@@ -137,6 +142,7 @@ export namespace CojsonInternalTypes {
   export type KnownStateMessage = import("./sync.js").KnownStateMessage;
   export type LoadMessage = import("./sync.js").LoadMessage;
   export type NewContentMessage = import("./sync.js").NewContentMessage;
+  export type SessionNewContent = import("./sync.js").SessionNewContent;
   export type CoValueHeader = import("./coValueCore.js").CoValueHeader;
   export type Transaction = import("./coValueCore.js").Transaction;
   export type TransactionID = import("./ids.js").TransactionID;