+++ tests/e2e/src/pages/FileStream/lib/getDownloaderPeerUrl.ts
@@ -1,0 +1,7 @@
+import { UploadedFile } from "../schema";
+
+export function getDownloaderPeerUrl(value: UploadedFile) {
+  const url = new URL(window.location.href);
+  url.searchParams.set("valueId", value.id);
+  return url.toString();
+}