+++ examples/onboarding/CHANGELOG.md
@@ -1,5 +1,20 @@
 # jazz-example-onboarding
 
+## 0.0.13
+
+### Patch Changes
+
+- jazz-react@0.8.34
+- jazz-tools@0.8.34
+- jazz-browser-media-images@0.8.34
+
+## 0.0.12
+
+### Patch Changes
+
+- jazz-browser-media-images@0.8.33
+- jazz-react@0.8.33
+
 ## 0.0.11
 
 ### Patch Changes