+++ examples/password-manager/CHANGELOG.md
@@ -1,5 +1,18 @@
 # jazz-password-manager
 
+## 0.0.31
+
+### Patch Changes
+
+- jazz-react@0.8.34
+- jazz-tools@0.8.34
+
+## 0.0.30
+
+### Patch Changes
+
+- jazz-react@0.8.33
+
 ## 0.0.29
 
 ### Patch Changes