+++ examples/pets/CHANGELOG.md
@@ -1,5 +1,20 @@
 # jazz-example-pets
 
+## 0.0.129
+
+### Patch Changes
+
+- jazz-react@0.8.34
+- jazz-tools@0.8.34
+- jazz-browser-media-images@0.8.34
+
+## 0.0.128
+
+### Patch Changes
+
+- jazz-browser-media-images@0.8.33
+- jazz-react@0.8.33
+
 ## 0.0.127
 
 ### Patch Changes