+++ packages/cojson-transport-ws/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cojson-transport-nodejs-ws
 
+## 0.8.34
+
+### Patch Changes
+
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+
 ## 0.8.32
 
 ### Patch Changes