+++ packages/jazz-browser-media-images/CHANGELOG.md
@@ -1,5 +1,19 @@
 # jazz-browser-media-images
 
+## 0.8.34
+
+### Patch Changes
+
+- jazz-browser@0.8.34
+- jazz-tools@0.8.34
+
+## 0.8.33
+
+### Patch Changes
+
+- Updated dependencies [3cb27e1]
+  - jazz-browser@0.8.33
+
 ## 0.8.32
 
 ### Patch Changes