+++ packages/jazz-run/CHANGELOG.md
@@ -1,5 +1,15 @@
 # jazz-run
 
+## 0.8.34
+
+### Patch Changes
+
+- Updated dependencies [e4f110f]
+  - cojson@0.8.34
+  - cojson-storage-sqlite@0.8.34
+  - cojson-transport-ws@0.8.34
+  - jazz-tools@0.8.34
+
 ## 0.8.32
 
 ### Patch Changes